% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivagraph.R
\name{vivagraph}
\alias{vivagraph}
\title{Vivagraph force layout}
\usage{
vivagraph(
  graph,
  layout = NULL,
  precompute_iterations = 1000,
  initial_size_multiplier = 75,
  pin_nodes = FALSE,
  pin_threshold = 4,
  pinned_cols = 2,
  pinned_rows = "auto",
  pinned_size_multiplier = 20,
  lcc_margin_left = 300
)
}
\arguments{
\item{graph}{igraph object}

\item{layout}{optional initial layout}

\item{precompute_iterations}{number of iterations for the magic precomputing step}

\item{initial_size_multiplier}{the precomputed layout size multiplier}

\item{pin_nodes}{whether to pin unconnected nodes to a grid}

\item{pin_threshold}{pin components with node count <= this number}

\item{pinned_cols}{number of columns to pin unconnected nodes}

\item{pinned_rows}{number of rows to pin unconnected nodes}

\item{pinned_size_multiplier}{pinned grid size}

\item{lcc_margin_left}{distance between the pinned grid and other nodes}
}
\value{
A 2 column layout matrix with node coordinates
}
\description{
Spawns a shiny app that generates a layout matrix with node coordinates
}
\examples{
g <- igraph::erdos.renyi.game(n = 5000, p.or.m = 10000, type = "gnm")

layout <- easylayout::vivagraph(g, pin_nodes = FALSE)
layout <- easylayout::vivagraph(g, layout = layout, pin_nodes = TRUE, pinned_cols = 10, lcc_margin_left = 500)

plot(g, layout = layout, vertex.size = 1, vertex.label = NA)

}
